dockerdesk/requirements.txt
Flask==2.0.0
mysql-connector-python==8.0.27
requests